+++ solidity/test/processor/ProcessorMessageDecoder.t.sol
@@ -2,8 +2,8 @@
 pragma solidity ^0.8.28;
 
 import {Test, console} from "forge-std/src/Test.sol";
-import {ProcessorMessageDecoder} from "../src/processor/libs/ProcessorMessageDecoder.sol";
-import {IProcessorMessageTypes} from "../src/processor/interfaces/IProcessorMessageTypes.sol";
+import {ProcessorMessageDecoder} from "../../src/processor/libs/ProcessorMessageDecoder.sol";
+import {IProcessorMessageTypes} from "../../src/processor/interfaces/IProcessorMessageTypes.sol";
 
 contract ProcessorMessageDecoderTest is Test {
     // Helper function to create a basic atomic subroutine
@@ -253,7 +253,7 @@ contract ProcessorMessageDecoderTest is Test {
         assertEq(atomicSubroutine.retryLogic.interval.value, 100);
 
         // Verify messages array contains data
-        for (uint256 i = 0; i < sendMsgs.messages.length; i++) {
+        for (uint8 i = 0; i < sendMsgs.messages.length; i++) {
             assert(sendMsgs.messages[i].length > 0);
         }
     }
